Angular6 Websocket demo
This is a simple project to show how we can easily connect to a websocket, send/receive data and close the websocket.
We pick up the publicly available websocket wss://echo.websocket.org which will just echo the input string. Basically
we are trying to build an Angular6 application similar to https://websocket.org/echo.html.